+++ data_handlers/serverbound_packets/1/distributer.cjs
@@ -4,7 +4,7 @@ const {Socket} = require('../../../data_structures.cjs')
  * @param {Socket} socket 
  */
 function ReadPacket(socket) {
-    if (socket.thisPlayer.upvn == 8) return require('./213.cjs').ReadPacket
+    if (socket.thisPlayer.upvn >= 8 && socket.thisPlayer.upvn <= 9) return require('./213.cjs').ReadPacket
     else {
         socket.log(`ERR: Cannot Parse Packet 1 for Version ${socket.thisPlayer.upvn}:${socket.thisPlayer.uvni}`)
         return () => {}