+++ data_handlers/serverbound_packets/15/213.cjs
@@ -138,7 +138,7 @@ function ReadPacket(world, socket, data) {
                                 }
                             } else socket.thisPlayer.tick.errorMessages.push("This block isn't available in all versions.")
                         } else if (utils.tag(world, placedName, "stairs") || placedName == "chest" || placedName == "furnace") {
-                            var playerDirection = utils.player.GetDirectionNESW(socket.thisPlayer.rotation.yaw)
+                            var playerDirection = utils.angle.GetDirectionNESW(socket.thisPlayer.rotation.yaw)
 
                             if (facingBlockReplacable) {
                                 if (validBlock) {
@@ -150,7 +150,7 @@ function ReadPacket(world, socket, data) {
                         } else if (placedName == "ladder") {
                             var playerDirection = "???"
 
-                            if (face.value == 0 || face.value == 1) playerDirection = utils.player.GetDirectionNESW(socket.thisPlayer.rotation.yaw + 180)
+                            if (face.value == 0 || face.value == 1) playerDirection = utils.angle.GetDirectionNESW(socket.thisPlayer.rotation.yaw + 180)
                             else {
                                 if (face.value == 2) playerDirection = "north"
                                 else if (face.value == 3) playerDirection = "south"
@@ -166,7 +166,7 @@ function ReadPacket(world, socket, data) {
                             } else socket.thisPlayer.tick.errorMessages.push("This block isn't available in all versions.")
                             giveItem = true
                         } else if (placedName == "rail") {
-                            var playerDirection = utils.player.GetDirectionNESW(socket.thisPlayer.rotation.yaw)
+                            var playerDirection = utils.angle.GetDirectionNESW(socket.thisPlayer.rotation.yaw)
 
                             if (playerDirection == "east" || playerDirection == "west") playerDirection = "east_west"
                             else playerDirection = "north_south"
@@ -180,7 +180,7 @@ function ReadPacket(world, socket, data) {
                             }
                         } else if (placedName == "lever") {
                             var playerDirectionFlipped = face.value != 1
-                            var playerDirection = utils.player.GetDirectionNESW(socket.thisPlayer.rotation.yaw + (playerDirectionFlipped ? 180 : 0))
+                            var playerDirection = utils.angle.GetDirectionNESW(socket.thisPlayer.rotation.yaw + (playerDirectionFlipped ? 180 : 0))
                             
                             var playerWall = "north"
                             if (face.value == 3) playerWall = "south"
@@ -199,7 +199,7 @@ function ReadPacket(world, socket, data) {
                                 giveItem = true
                             }
                         } else if (placedName == "stone_button") {
-                            var playerDirection = utils.player.GetDirectionNESW(socket.thisPlayer.rotation.yaw + 180)
+                            var playerDirection = utils.angle.GetDirectionNESW(socket.thisPlayer.rotation.yaw + 180)
                             
                             var playerWall = "north"
                             if (face.value == 3) playerWall = "south"
@@ -280,7 +280,7 @@ function ReadPacket(world, socket, data) {
                                         utils.tick_actions.set_block.AddBlockUpdate(socket)(world, socket, facingBlock, `oak_wall_sign[facing=${playerDirection}]`, true, facingBlock.block)
                                         updateSuccessful = true
                                     } else {
-                                        var playerDirection = utils.player.GetDirection16Num(socket.thisPlayer.rotation.yaw + 180)
+                                        var playerDirection = utils.angle.GetDirection16Num(socket.thisPlayer.rotation.yaw + 180)
 
                                         utils.tick_actions.set_block.AddBlockUpdate(socket)(world, socket, facingBlock, `oak_sign[rotation=${playerDirection}]`, true, facingBlock.block)
                                         updateSuccessful = true
@@ -295,7 +295,7 @@ function ReadPacket(world, socket, data) {
                                     if (utils.tag(world, aboveFacingBlock, "replaceable")) {
                                         giveItem = true
                                         if (facingBlock.y < 63) {
-                                            var playerFacing = utils.player.GetDirectionNESW(socket.thisPlayer.rotation.yaw)
+                                            var playerFacing = utils.angle.GetDirectionNESW(socket.thisPlayer.rotation.yaw)
 
                                             var connectingDoorBlock = "air"
                                             if (playerFacing == "north") connectingDoorBlock = utils.worldgen.GetBlock(socket)(world, socket, {x: facingBlock.x - 1, y: facingBlock.y, z: facingBlock.z})