+++ CHANGELOG.md
@@ -4,6 +4,34 @@ This is a general overview of how the server protocol changes from version to ve
 
 For more information, view the [development spreadsheet](https://docs.google.com/spreadsheets/d/1rxPJ9sm80nxbCEAkAXiAbzHODu7CdnCulBLebjD1Lhk/edit?usp=sharing).
 
+## Java Edition Alpha v1.0.16_02
+
+This version can't be distinguished from v1.0.16.
+
+### Documentation
+
+Protocol documented on the [Minecraft Wiki](https://minecraft.wiki/w/Minecraft_Wiki:Projects/wiki.vg_merge/Protocol?oldid=2769659)
+
+## Java Edition Alpha v1.0.16_01
+
+This version can't be distinguished from v1.0.16.
+
+### Documentation
+
+Protocol documented on the [Minecraft Wiki](https://minecraft.wiki/w/Minecraft_Wiki:Projects/wiki.vg_merge/Protocol?oldid=2769659)
+
+## Java Edition Alpha v1.0.16
+
+Protocol Version set to 14.
+
+### Packets
+
+Added Handshake clientbound and serverbound packet for account authentication (obsolete).
+
+### Documentation
+
+Protocol documented on the [Minecraft Wiki](https://minecraft.wiki/w/Minecraft_Wiki:Projects/wiki.vg_merge/Protocol?oldid=2769659)
+
 ## Java Edition Alpha v1.0.15
 
 Protocol Version set to 13.